+++ src/pages/Main/Main.tsx
@@ -1,11 +1,51 @@
 import React from 'react';
 import Quote from '../../components/Quote/Quote';
-import styled from 'styled-components';
+import styled, { css, useTheme } from 'styled-components';
 import { QuoteType } from 'components/Quote/Quote.type';
+import Feed from 'components/Feed/Feed';
+import { useTranslation } from 'react-i18next';
+import Switcher from 'components/Switcher/Switcher';
+import GuideLinks from 'components/GuideLinks/GuideLinks';
+import Markdown from 'react-markdown';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import Badge from 'components/Badge/Badge';
+import { BadgeStyles } from 'components/Badge/Badge.type';
 
 const MainContainer = styled.div`
   display: grid;
-  gap: ${({ theme }) => theme.spacing.s.rem};
+  grid-template-areas: 
+    // 'users users'
+    'quotes feeds';
+  grid-template-columns: 1fr auto;
+
+  ${({ theme }) => `
+    gap: ${theme.spacing.s.rem};
+
+    @media (max-width: ${theme.breakpoints.lg}) {
+      grid-template-areas: 'quotes';
+      grid-template-columns: 1fr;
+    }
+  `}
+`;
+
+const Container = css`
+  display: flex;
+  flex-direction: column;
+  gap: inherit;
+`;
+
+const QuotesContainer = styled.div`
+  ${Container}
+  grid-area: quotes;
+`;
+
+const FeedsContainer = styled.div`
+  position: sticky;
+  max-width: 400px;
+  ${Container}
+  grid-area: feeds;
+  top: 0;
+  place-self: start;
 `;
 
 // TODO: Default Icons for the Quotes, replace it in the future
@@ -104,11 +144,86 @@ const quotes: QuoteType[] = [
  * Main Page for Quotly
  */
 const Main = () => {
+  const { t } = useTranslation();
+  const theme = useTheme();
+
   return (
     <MainContainer>
-      {quotes.map((quote, index) => (
-        <Quote {...quote} key={index} isLast={quotes.length == (index+1)} />
-      ))}
+      <QuotesContainer>
+        {quotes.map((quote, index) => (
+          <Quote {...quote} key={index} isLast={quotes.length == (index+1)} />
+        ))}
+      </QuotesContainer>
+      <Switcher
+        breakpoint={theme.breakpoints.lg}
+        desktop={
+          <FeedsContainer>
+            <Feed title={t('feeds.top_quotes')} items={
+              quotes.slice(0, 3).map((quote, index) => {
+                const colors = [theme.colors.gold, theme.colors.silver, theme.colors.bronze];
+
+                return {
+                  item: (
+                    <React.Fragment>
+                      <Badge 
+                        style={BadgeStyles.custom}
+                        color={colors[index]}
+                        fontSize={theme.font.sizes.xs.rem}
+                        children={<FontAwesomeIcon icon={(index + 1).toString() as '1' | '2' | '3'} />} 
+                      /> 
+                      <Markdown children={quote.quote.text} />
+                    </React.Fragment>
+                  ),
+                  url: quote.quote.url
+                };
+              })
+            }/>
+            <Feed title={t('feeds.suggested_profiles')} items={
+              [
+                {
+                  item: 'Daniel',
+                  url: '/'
+                },
+                {
+                  item: 'Domi',
+                  url: '/'
+                },
+                {
+                  item: 'Rubinschwein47',
+                  url: '/'
+                },
+                {
+                  item: 'Jordan',
+                  url: '/'
+                },
+                {
+                  item: 'Daniel',
+                  url: '/'
+                },
+                {
+                  item: 'Domi',
+                  url: '/'
+                },
+                {
+                  item: 'Rubinschwein47',
+                  url: '/'
+                },
+                {
+                  item: 'Jordan',
+                  url: '/'
+                }
+              ].slice(0, 5)
+            } />
+            <GuideLinks
+              links={[
+                { label: t('guides.privacy_policy'), url: '/privacy' },
+                { label: t('guides.terms_of_service'), url: '/tos' }
+              ]}
+            />
+          </FeedsContainer>
+        }
+        mobile={<></>}
+      />
     </MainContainer>
   );
 };